Trait moonbeam_finality_rpc::MoonbeamFinalityApi

pub trait MoonbeamFinalityApi: Sized + Send + Sync + 'static {
    fn is_block_finalized(
        &self, 
        block_hash: H256
    ) -> BoxFuture<'static, RpcResult<bool>>;

    fn is_tx_finalized(
        &self, 
        tx_hash: H256
    ) -> BoxFuture<'static, RpcResult<bool>>;

    fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M> { ... }
}

An RPC endpoint to check for finality of blocks and transactions in Moonbeam

Required methods

fn is_block_finalized(
&self,
block_hash: H256
) -> BoxFuture<'static, RpcResult<bool>>

Reports whether a Substrate or Ethereum block is finalized. Returns false if the block is not found.

fn is_tx_finalized(&self, tx_hash: H256) -> BoxFuture<'static, RpcResult<bool>>

Reports whether an Ethereum transaction is finalized. Returns false if the transaction is not found

Provided methods

fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M>

Create an IoDelegate, wiring rpc calls to the trait methods.