Trait moonbeam_finality_rpc::MoonbeamFinalityApiServer

pub trait MoonbeamFinalityApiServer: Sized + Send + Sync + 'static {
    fn is_block_finalized(&self, block_hash: H256) -> RpcResult<bool>;
    fn is_tx_finalized(&self, tx_hash: H256) -> RpcResult<bool>;

    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the MoonbeamFinalityApi RPC API.

Required Methods

fn is_block_finalized(&self, block_hash: H256) -> RpcResult<bool>

Reports whether a Substrate or Ethereum block is finalized. Returns false if the block is not found.

fn is_tx_finalized(&self, tx_hash: H256) -> RpcResult<bool>

Reports whether an Ethereum transaction is finalized. Returns false if the transaction is not found

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.